pyproject.toml
[tool.poetry]
name = "ahc-app"
version = "0.1.0"
description = ""
authors = ["notkamonohasi <notkamonohasi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"
boto3 = "^1.34.129"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
types-requests = "^2.32.0.20240602"


[tool.poetry.group.dev.dependencies]
boto3-stubs = {version = "1.34.129", extras = ["lambda", "s3", "ssm"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
